chores

+++ kernel/defs.h
@@ -173,7 +173,7 @@ int             copyout(pagetable_t, uint64, char *, uint64);
 int             copyin(pagetable_t, char *, uint64, uint64);
 int             copyinstr(pagetable_t, char *, uint64, uint64);
 int             cow_check(pagetable_t, uint64);
-int             cow_copy(pagetable_t, uint64);
+uint64          cow_copy(pagetable_t, uint64);
 
 // plic.c
 void            plicinit(void);
